+++ src/TooMany.Actors/Worker/Processes/ProcessSupervisor.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using System.IO;
 
 namespace TooMany.Actors.Worker.Processes
 {