+++ src/TooMany.Cli/Commands/StopTaskCommand.cs
@@ -19,15 +19,15 @@ namespace TooMany.Cli.Commands
 			ShowUnknownOptions(context);
 			ShowIgnoredArguments(context);
 
-			var tasks = await GetTasks(settings).WaitWith("Getting task list...");
+			var tasks = await GetTasks(settings).WithSpinner("Getting task list...");
 
 			if (tasks.Length <= 0) return 0;
 
 			var found = tasks.Select(t => t.Name).ToArray();
 
-			await Task.WhenAll(found.Select(n => Host.StopTask(n))).WaitWith("Stopping tasks...");
+			await Task.WhenAll(found.Select(n => Host.StopTask(n))).WithSpinner("Stopping tasks...");
 
-			var refreshed = await GetNamedTasks(found).WaitWith("Refreshing task states...");
+			var refreshed = await GetNamedTasks(found).WithSpinner("Refreshing task states...");
 
 			Presentation.TaskInfo(refreshed);